rename utilities to util
