+++ mavehgvs/__init__.py
@@ -1,6 +1,6 @@
 from mavehgvs.exceptions import MaveHgvsParseError
 from mavehgvs.position import VariantPosition
 from mavehgvs.variant import Variant
-from mavehgvs.utilities import parse_variant_strings
+from mavehgvs.util import parse_variant_strings
 
 __all__ = ["Variant", "VariantPosition", "MaveHgvsParseError", "parse_variant_strings"]
